Changed buyIn type to double
This change is required for importing google docs tabel

+++ src/main/java/de/stammtisch/pokerstats/models/UserGame.java
@@ -22,5 +22,5 @@ public class UserGame {
     @ManyToOne(fetch = FetchType.EAGER)
     private PokerGame pokerGame;
 
-    private int buyIn;
+    private double buyIn;
 }
